requirements.txt
python-telegram-bot[job-queue]==20.7
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.0
dateparser==1.2.0
pytz==2024.1
